backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
passlib[bcrypt]
python-multipart
python-jose[cryptography]
faker
